Service and Dao 통합 및 단축 하고 테스트 완료
