--- jPanda/src/main/java/com/kakao/jPanda/lhw/dao/NoticeDao.java
@@ -1,10 +1,0 @@
-package com.kakao.jPanda.lhw.dao;
-
-import java.util.List;
-
-import com.kakao.jPanda.lhw.domain.Notice;
-
-public interface NoticeDao {
-
-	List<Notice> selectNoticeList();
-}
