+++ jPanda/src/main/java/com/kakao/jPanda/lhw/dao/TalentDao.java
@@ -2,17 +2,14 @@ package com.kakao.jPanda.lhw.dao;
 
 import java.util.List;
 
+import com.kakao.jPanda.lhw.domain.Review;
 import com.kakao.jPanda.lhw.domain.Talent;
 
 public interface TalentDao {
 
-	List<Talent> selectTalentList();
-
+	// 재능 판매 상세 페이지 불러오기
 	Talent selectTalentByTalentNo(Long talentNo);
-
-	List<Talent> selectTalentListByUpperCategoryNo(Long upperCategoryNo);
-
-	List<Talent> selectTalentListByLowerCategoryNo(Long lowerCategoryNo);
 	
-	
+	// 리뷰 리스트 불러오기
+	List<Review> selectReivewListByTalentNo(Long talentNo);
 }
