+++ jPanda/src/main/java/com/kakao/jPanda/lhw/dao/TalentDaoImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.stereotype.Repository;
 
+import com.kakao.jPanda.lhw.domain.Review;
 import com.kakao.jPanda.lhw.domain.Talent;
 
 import lombok.RequiredArgsConstructor;
@@ -12,28 +13,20 @@ import lombok.RequiredArgsConstructor;
 @Repository
 @RequiredArgsConstructor
 public class TalentDaoImpl implements TalentDao {
-
+	
 	private final SqlSession sqlSession;
 	
-	@Override
-	public List<Talent> selectTalentList() {
-		return sqlSession.selectList("selectTalentList");
-	}
-	
+	// 재능 판매 상세 페이지 불러오기
 	@Override
 	public Talent selectTalentByTalentNo(Long talentNo) {
 		return sqlSession.selectOne("selectTalentByTalentNo", talentNo);
 	}
-
+	
+	// 리뷰 리스트 불러오기
 	@Override
-	public List<Talent> selectTalentListByUpperCategoryNo(Long upperCategoryNo) {
-		return sqlSession.selectList("selectTalentListByUpperCategoryNo", upperCategoryNo);
+	public List<Review> selectReivewListByTalentNo(Long talentNo) {
+		return sqlSession.selectList("selectReviewListByTalentNo", talentNo);
 	}
-
-	@Override
-	public List<Talent> selectTalentListByLowerCategoryNo(Long lowerCategoryNo) {
-		return sqlSession.selectList("selectTalentListByLowerCategoryNo", lowerCategoryNo);
-	}
-
+	
 
 }
