--- jPanda/src/main/java/com/kakao/jPanda/lhw/service/ReviewService.java
@@ -1,11 +1,0 @@
-package com.kakao.jPanda.lhw.service;
-
-import java.util.List;
-
-import com.kakao.jPanda.lhw.domain.Review;
-
-public interface ReviewService {
-	
-	List<Review> findReviewListByTalentNo(Long talentNo);
-	
-}
